+++ src/compiler/asm/AsmBuilder.java
@@ -471,13 +471,13 @@ public class AsmBuilder {
         return regOperand(RegOperandOP.MOV, rd, new RegOperand(rn));
     }
 
-    public AsmBuilder mov(Reg rd, long imm8m) {
+    public AsmBuilder mov(Reg rd, int imm8m) {
         if(_hookIfNotImmXX)
         {
             if(!AsmUtil.imm8m(imm8m))
             {
                 //ldrEq(rd, imm8m);
-                dataHolder.addAndLoadFromFuncData(this,(int) imm8m,rd);
+                dataHolder.addAndLoadFromFuncData(this,imm8m,rd);
                 return this;
             }
         }
@@ -492,7 +492,7 @@ public class AsmBuilder {
         return regRegOperand(RegRegOperandOP.ADD, rd, rn, new RegOperand(rm));
     }
 
-    public AsmBuilder add(Reg rd, Reg rn, long imm8m) {
+    public AsmBuilder add(Reg rd, Reg rn, int imm8m) {
         if(_hookIfNotImmXX)
         {
             if(!AsmUtil.imm8m(imm8m))
@@ -503,7 +503,7 @@ public class AsmBuilder {
                 else
                     tmp = regGetter.getTmpRegister();
                 //ldrEq(tmp, imm8m);
-                dataHolder.addAndLoadFromFuncData(this,(int)imm8m,tmp);
+                dataHolder.addAndLoadFromFuncData(this,imm8m,tmp);
                 return add(rd,rn,tmp);
             }
         }
@@ -519,7 +519,7 @@ public class AsmBuilder {
         return regRegOperand(RegRegOperandOP.SUB, rd, rn, new RegOperand(rm));
     }
 
-    public AsmBuilder sub(Reg rd, Reg rn, long imm8m) {
+    public AsmBuilder sub(Reg rd, Reg rn, int imm8m) {
         if(_hookIfNotImmXX)
         {
             if(!AsmUtil.imm8m(imm8m))
@@ -530,7 +530,7 @@ public class AsmBuilder {
                 else
                     tmp = regGetter.getTmpRegister();
                 //ldrEq(tmp, imm8m);
-                dataHolder.addAndLoadFromFuncData(this,(int) imm8m,tmp);
+                dataHolder.addAndLoadFromFuncData(this,imm8m,tmp);
                 return sub(rd,rn,tmp);
             }
         }
