+++ src/compiler/genir/code/AddressOrData.java
@@ -3,13 +3,13 @@ package compiler.genir.code;
 import compiler.asm.Reg;
 
 public class AddressOrData {
-    public AddressOrData(boolean isData, long item) {
+    public AddressOrData(boolean isData, int item) {
         this.isData = isData;
         this.item = item;
     }
 
     public boolean isData;
-    public long item;
+    public int item;
     //public Reg reg;
     @Override
     public String toString() {
